+++ src/helper_functions.py
@@ -1,6 +1,23 @@
 import pandas as pd
 from sklearn.model_selection import train_test_split
 from typing import Tuple
+import tf_keras as tfk
+import tqdm
+
+
+class EpochProgressBar(tfk.callbacks.Callback):
+    def on_train_begin(self, logs=None):
+        self.epochs = self.params.get("epochs", 1)
+        self.progress_bar = tqdm(
+            total=self.epochs, desc="Training Progress", unit="epoch"
+        )
+
+    def on_epoch_end(self, epoch, logs=None):
+        self.progress_bar.update(1)
+
+    def on_train_end(self, logs=None):
+        self.progress_bar.close()
+
 
 def get_train_test(
     data: pd.DataFrame, test_size: int = 2, random_state: int = None
